Add simple endpoint tests, start react component testing
