+++ __tests__/index.test.js
@@ -1,0 +1,13 @@
+// import React from 'react';
+// import { render, cleanup } from 'react-testing-library';
+// import '@testing-library/jest-dom';
+// import Component from '../client/componentName';
+
+// to be completed after add-components is approved & merged with master
+
+// test.afterEach(cleanup);
+
+// test('renders', () => {
+//   const { asFragment } = render(<Component prop="whatever" />);
+//   expect(asFragment()).toMatchSnapshot();
+// });
